fix bug (Mark-up does not reflect updated terminology)
When then Terminology page is edited, the new revision is not available in the database
yet when the glossary is built from it. This leads to the changes not being reflected
in the markup of the Terminology page.

The fix is to in this special case not read the glossary from then Terminology page in
the database, but take it from the text submitted in the web request.

Change-Id: Iaf4d5e1476d1df5c681b2d1675947973515353be