Add support for text transformation API parser options
Some ApiParse params (disabletoc, disableeditsection, wrapoutputclass, disablestylededuplication) are not set in the ParserOptions object in the ApiParse class
These Parser options are not passed into Lingo Extension when the 'ContentAlterParserOutput' Hook is executed.
When Lingo execute the Parser call, it did not have a full set of options to completely process the text according to all the originally set api parameters.

Bug: T219109
Change-Id: I92d37be5bd527b9c2f50d08999a6ab37d70d6e6e