Simplify extension description message
This fixes the extension description message on MW 1.35+ *and* fixes the linked bug by removing the offending code.
Yes, it has the side effect of users not immediately being able to tell whether the Lingo page is the one defined
on [[MediaWiki:Lingo-terminologypagename]] or the one set in LocalSettings.php as $wgexLingoPage, but it's an
acceptable trade-off in my opinion.

Bug: T283604
Change-Id: I9f7d4070beea68fc4d153d945f9d32da9eaffdd1