Fix case where we fail to match a long word and forget about prefix
If you have a definition that is multiple lexmes, as well as one
that is a prefix, previously if it matched in between it would
match nothing instead of just the prefix.

e.g. If you have two definitions, "foo" and "foo bar". If you
have "foo baz" in the page, it would match up to "foo ", realize
it can't match any further, try to return the definition of "foo ",
of which there is none. Instead we need to back track to the valid
match "foo".

Reported https://www.mediawiki.org/wiki/Topic:Xb96lwtzsrmne98s [WE-239]

Change-Id: I49efd3c0c5ad5055e29c0e28a5ad7de9b2a69b15