Ensure that LingoParser does not run when fillParserOutput is false
This prevents bugs when used with Extension:SpamBlacklist.

This fixes the immediate issue, but the assumptions this extension
makes about the MW parser in general are unsound with content
handler and should eventually be rewritten. Namely, that the global
instance of the Parser is not neccesarily the same instance of the
parser that the render that just happened is using. For that matter,
content handlers are not obliged to use the MediaWiki parser at all.

Change-Id: I45c58ecccc18e742941251f3af6689c72a1908b2