Remove smaller pieces of dead, unused code
Notably, there is no need to check if the incoming Parser really is
a Parser or a stub. The stub is a subclass of the original Parser
class.

Having test logic bleed into production code is not a good idea, and
can easily be avided in this case here, I believe.

Change-Id: I95c76072eaef93304af48f582dc1779a340b88ab